feat(frontend): remove code duplication
